Remove Counter init function from ex 2, 3 & 4 to match ex 1
Example 2 explains how the Counter created in example 1 can be reused
without change. Currently this isn't quite true as the Counter used in
example 2, 3 & 4 all define an additional init function not defined in
example 1. This commit removes the init function in favour of parent
components explicitly passing an Int to represent the Counter Model